+++ extensions/forked-extensions/src/manage-forked-extensions.tsx
@@ -18,7 +18,7 @@ import { catchError, handleError } from "./errors.js";
 import * as git from "./git.js";
 import operation from "./operation.js";
 import { ForkedExtension } from "./types.js";
-import { extensionLink, getActualIconPath, userLink } from "./utils.js";
+import { extensionLink, getActualIconPath, userLink, openWith } from "./utils.js";
 
 function ManageForkedExtensions() {
   const [isShowingDetail, setIsShowingDetail] = useState(false);
@@ -106,6 +106,14 @@ function ManageForkedExtensions() {
             <ActionPanel>
               <ActionPanel.Section>
                 <Action icon={Icon.Eye} title="Show Details" onAction={() => setIsShowingDetail(!isShowingDetail)} />
+                {openWith ? (
+                  <Action.Open
+                    application={openWith.bundleId}
+                    target={x.folderPath}
+                    title={`Open with ${openWith.name}`}
+                    icon={{ fileIcon: openWith.path }}
+                  />
+                ) : null}
                 <Action.OpenWith path={x.folderPath} />
                 <Action.CopyToClipboard
                   title="Copy Extension Path to Clipboard"